Always request rc to report absolute paths
With Andersbakken/rtags#429 the default behavior of `rc` was changed so
that now all reported paths are by default relative to the project root.
To accomodate this we should always explicitly request absolute paths
with `--absolute-path`. This flag has existed for quite a time so I
expect little breakage when interfacing older rtags versions.