Only send unsaved content to rdm if modified.
If the current buffer is a header file on which many targets depend, then
whenever invoking a follow or jump command, the entire file was sent to
rdm, whether modified or not, which triggered a large recompile.  As a
result, it could take a while for the header to become indexed, so the
jump failed.

This change optimises the sending of unsaved content to rdm, to only send
it if the buffer 'modified' flag is set, or if any unsaved content has
already been sent from the same buffer, and the buffer content has changed
in the meantime.